Fixed Code smells
